ISPN-8074 testComputeIfPresentFromNonOwner correction
