+++ core/src/test/java/org/infinispan/distribution/DistSyncFuncTest.java
@@ -9,6 +9,7 @@ import static org.testng.AssertJUnit.assertTrue;
 
 import java.util.ArrayList;
 import java.util.Collection;
+import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 import java.util.Random;
@@ -20,6 +21,7 @@ import org.infinispan.commands.functional.ReadWriteKeyCommand;
 import org.infinispan.commands.write.ClearCommand;
 import org.infinispan.commands.write.ComputeCommand;
 import org.infinispan.commands.write.ComputeIfAbsentCommand;
+import org.infinispan.commands.write.InvalidateL1Command;
 import org.infinispan.commands.write.PutKeyValueCommand;
 import org.infinispan.commands.write.RemoveCommand;
 import org.infinispan.commands.write.ReplaceCommand;
@@ -27,6 +29,7 @@ import org.infinispan.commons.util.ObjectDuplicator;
 import org.infinispan.context.Flag;
 import org.infinispan.remoting.RemoteException;
 import org.infinispan.remoting.transport.Address;
+import org.infinispan.test.ReplListener;
 import org.infinispan.test.TestingUtil;
 import org.infinispan.util.function.SerializableBiFunction;
 import org.infinispan.util.function.SerializableFunction;
@@ -37,6 +40,32 @@ public class DistSyncFuncTest extends BaseDistFunctionalTest<Object, String> {
 
    public DistSyncFuncTest() {
       testRetVals = true;
+   }
+
+   protected Map<Cache<?, ?>, ReplListener> listenerLookup;
+
+   protected ReplListener r1, r2, r3, r4;
+   protected ReplListener[] r;
+
+   @Override
+   protected void createCacheManagers() throws Throwable {
+      super.createCacheManagers();
+      r1 = new ReplListener(c1, true, true);
+      r2 = new ReplListener(c2, true, true);
+      r3 = new ReplListener(c3, true, true);
+      r4 = new ReplListener(c4, true, true);
+      r = new ReplListener[]{r1, r2, r3, r4};
+      listenerLookup = new HashMap<>();
+      for (ReplListener rl : r) listenerLookup.put(rl.getCache(), rl);
+   }
+
+   private void waitForInvalidateL1InNonOwners(String key) {
+      if(l1CacheEnabled) {
+         for (Cache<?, ?> c : getNonOwners(key)) {
+            listenerLookup.get(c).expect(InvalidateL1Command.class);
+            listenerLookup.get(c).waitForRpc();
+         }
+      }
    }
 
    public void testLocationConsensus() {
@@ -64,9 +93,9 @@ public class DistSyncFuncTest extends BaseDistFunctionalTest<Object, String> {
       List l3 = getCacheTopology(c3).getDistribution(key).writeOwners();
       List l4 = getCacheTopology(c4).getDistribution(key).writeOwners();
 
-      assertEquals("L1 "+l1+" and L2 "+l2+" don't agree.", l1, l2);
-      assertEquals("L2 "+l2+" and L3 "+l3+" don't agree.", l2, l3);
-      assertEquals("L3 "+l3+" and L4 "+l4+" don't agree.", l3, l4);
+      assertEquals("L1 " + l1 + " and L2 " + l2 + " don't agree.", l1, l2);
+      assertEquals("L2 " + l2 + " and L3 " + l3 + " don't agree.", l2, l3);
+      assertEquals("L3 " + l3 + " and L4 " + l4 + " don't agree.", l3, l4);
 
    }
 
@@ -343,6 +372,7 @@ public class DistSyncFuncTest extends BaseDistFunctionalTest<Object, String> {
       Object retval = getFirstNonOwner("k1").computeIfPresent("k1", (k, v) -> "computed_" + k + "_" + v);
       if (testRetVals) assertEquals("computed_k1_value", retval);
       asyncWait("k1", ComputeCommand.class);
+      waitForInvalidateL1InNonOwners("k1");
       assertOnAllCachesAndOwnership("k1", "computed_k1_value");
 
       RuntimeException computeRaisedException = new RuntimeException("hi there");
@@ -354,6 +384,7 @@ public class DistSyncFuncTest extends BaseDistFunctionalTest<Object, String> {
       // remove if after compute value is null
       retval = getFirstNonOwner("k1").computeIfPresent("k1", (v1, v2) -> null);
       asyncWait("k1", ComputeCommand.class);
+      waitForInvalidateL1InNonOwners("k1");
       if (testRetVals) assertNull(retval);
       assertRemovedOnAllCaches("k1");
